fix: allow OVO code URIs as competent and executing authorities
Product instances can have a competent and/or executing authority which is not
an administrative unit. Instead these are resources simply of type
`skos:Concept` and identified using a URI with their OVO code. Initially the
publishing validation only tried to retrieve administrative unit resources for
any competent and executing authorities. This resulted in incorrect error
messages when a user selected an organisation identified by an OVO code URI.

These changes extend the validation logic to correctly deal with
non-administrative unit competent and executing authorities. The validation
service now performs different checks based on the shape of the a URI. For OVO
code URIs the `CodeRepository` is used to check whether a concept for that URI
exists. For other URI the behaviour remains the same as before.