chore: fix failing test
The behaviour of the tested function was slightly changed in #45. Since then
competent authorities are only taken from the concept if the instance to update
does **not** have local as competent authority level. The test still assumed the
old behaviour in which competent authorities are always taken from the concept.

This fixes the test by correcting the `expectedInstance` to have the same
competent authorities of the updated instance. In the test at hand the competent
authorities of the instance are not changed during update since is has a local
competent authority level.