fix: run IPDC tests against correct API
Run IPDC integration tests against the actual IPDC TNI API instead of accessing
it via the web URL, which no longer works due to changes at IPDC. The TNI API
does require authentication via an API key. This key is specified in a `.env`
file in the `lpdc-management-ci` project. The package `dotenv` is added as a dev
dependency to support passing this key to the jest tests.