fix: replace NUTS URI placeholder in concept characteristics form
As part of the NUTS/LAU code list update its concept scheme URI was extracted to
a constant to avoid it being specified in multiple places. This required putting
a placeholder string in the characteristics form file, which is overwritten with
the actual value when the form file is loaded.

For concept forms the step to overwrite the place holder was not added resulting
in a form that is loaded with the placeholder string stil in it. Since concept
forms are not editable in LPDC this currently does not lead to any issues. This
commit does add the replacement for the placeholder string such that a correct
form is always sent to the frontend. This makes it consistent with other
placeholders that are already replaced and prevents (subtle) bugs should the
concept forms ever become editable.