chore: fix tests for transfer instance script
In #70 and #71 the transfer instance script was extended with new
functionality. But the PRs it neglected to update the tests for the script
accordingly. The caused running this service's test suite, using
`tests/run-test.sh`, to fail.

This commit updates the tests in question so that they run again. No new tests
to cover the scripts new functionality were added.